boot.js: prevent 'baselayerchange' to fire second time (on mobile)
Fix issue with minimap.

Background: mobile app uses native layer chooser instead of standard Control.Layers,
so window.layerChooser has additional code to support that.

Perhaps it was needed to fire 'baselayerchange' manually a long time ago,
but not in leaflet 1.

!!This is just quick fix for particular issue appeared with minimap on mobile!!
But in fact the whole layerChooser should be reimplemented from scratch.
Mobile-specific parts of code should be separated.