map_data_render.js: placeholders get -1 timestamp value
This can fix issue with recent intel bug, when normal portals (not placeholders) had 0 timestamp value,
which lead to wrong data processing, because detailed data was rejected as it considered not fresher than
placeholders (that originated from links/fields that currently are processing earlier).
https://t.me/iitc_group/8046

Note: this is rather temporary fix, as we need to revise our data procesing approaches in whole.
Perhaps, instead of relying on timestamps, we should compare new data details with cached,
every single property.

There is another issue: ghosted fields, which may be considered as intel (or Ingress) bug,
but we still should try to investigate it further, as it may be possible to propose some workaround.
Of cause if it will not turn out that Ingress itself is affected in same way (shows same ghosted fields,
and prevents linking in that area).

For now, just add more logging.