Use S2 coordinate system (again)
S2 is used on Ingress backend. By using the S2 Earth-radius constant we
fix the distance calculation.
User were reporting that DrawTools link-length didn't match the distance
shown in the ingress client  (of by ~0.2%)

We need to change L.CRS.Earth.R value for standard LatLng.distanceTo
calculation. And we need to setup a CRS for the Leaflet draw plugin.