boot.js: fix latitude clamping on map panning
There was an exception when panning beyond N || S bounds of the map:
`Maximum call stack size exceeded`, caused by looping of our clamping code.
This was introduced in iitc-ce by some internal leaflet changes.

So now for the same purpose we use standard leaflet facility, which is a way better,
because it limits the whole map bounds (not only map center as was before).