Automatic sending of IITC-Mobile app to Google Play (#305)
Added build process for the main branch and automatic submission of the IITC-Mobile app to Google Play in beta status.
Refusing to use number of commits as `versionCode` due to possibility of reducing the number of `versionCode`, which requires reinstalling android application. Using a date of commit as versionCode.
Enabled `proguard` rules for beta and release builds.