excludedOrnaments
excludedOrnaments [] holds a list of ornaments that  shall not be displayed, regardless of layer settings.
This is beneficial, if during an anomaly we want to see RBBs/VRBBs, but no victory ornaments.
Or if NIA introduces new types of ornaments that the user wants to hide, while IITC has not been updated to handle them.