basemap-kartverket: Update to new WMTS server
Statkart's opencache server is deprecated and will be phased out. It is already experiencing performance issues that makes it more or less useless as a basemap tileserver. The plugin is updated to use the new (and much faster) WMTS cache from Kartverket, as described here: https://kartverket.no/om-kartverket/nyheter/alle/2024/juli/treghet-pa-opencache.statkart.no

Many of the map layers that were available on opencache are not present on the new cache, and reasonable mappings have been set to try to mitigate this.