Expand `makePrimeLink()` into the component parts.
There is no functional difference with this change; it just builds the URL on
the fly.

Firebase is sunsetting this feature:
https://firebase.google.com/support/dynamic-links-faq

It is unknown if link.ingress.com still uses Firebase, or has implemented its
own work-alike.  But by documenting it this way, it may help any future
changes that come along.

Also, it opens up the possibility to customize links.  For example, I use a
similar breakdown in a Python to create deep links that leaves off the `ios`
portion, since I personally do not need them, and they are much shorter.